+++ src/main/java/com/buuz135/industrial/jei/manual/ManualWrapper.java
@@ -23,11 +23,28 @@ public class ManualWrapper implements IRecipeWrapper {
     private final TextureButton button;
     @Getter
     private final CategoryEntry entry;
+    private PageText text;
+    private int extraLines;
 
     @SideOnly(Side.CLIENT)
     public ManualWrapper(CategoryEntry entry) {
-        this.button = new TextureButton(-135, 0, 150, 18, 18, GUIBookMain.BOOK_EXTRAS, 1, 38, "Open Manual Entry");
+        this.button = new TextureButton(-135, 0, 105, 18, 18, GUIBookMain.BOOK_EXTRAS, 1, 38, "Open Manual Entry");
         this.entry = entry;
+        this.extraLines = 0;
+        int index = 0;
+        for (IPage page : entry.getPages()) {
+            if (page instanceof PageText) {
+                this.text = (PageText) page;
+                break;
+            }
+        }
+        while (index != -1) {
+            index = text.getText().indexOf("\n", index);
+            if (index != -1) {
+                extraLines++;
+                index += "\n".length();
+            }
+        }
     }
 
     @Override
@@ -42,13 +59,8 @@ public class ManualWrapper implements IRecipeWrapper {
         button.drawButton(minecraft, mouseX, mouseY, 0);
         int y = 20;
         minecraft.fontRenderer.drawString(TextFormatting.DARK_AQUA + entry.getName(), 0, y, 0xFFFFFF);
-        for (IPage page : entry.getPages()) {
-            if (page instanceof PageText) {
-                PageText pageText = (PageText) page;
-                minecraft.fontRenderer.drawSplitString((TextFormatting.DARK_GRAY + pageText.getText().substring(0, Math.min(250, pageText.getText().length())) + (pageText.getText().length() > 250 ? "..." : "")).replaceAll(TextFormatting.GOLD.toString(), TextFormatting.DARK_AQUA.toString()), 0, y + minecraft.fontRenderer.FONT_HEIGHT + 4, 160, 0xFFFFFF);
-                break;
-            }
-        }
+        minecraft.fontRenderer.drawSplitString((TextFormatting.DARK_GRAY + text.getText().substring(0, Math.min(200 - extraLines * 10, text.getText().length())) + (text.getText().length() > 200 - extraLines * 10 ? "..." : "")).replaceAll(TextFormatting.GOLD.toString(), TextFormatting.DARK_AQUA.toString()), 0, y + minecraft.fontRenderer.FONT_HEIGHT + 4, 160, 0xFFFFFF);
+
     }
 
     @SideOnly(Side.CLIENT)