+++ src/main/java/com/buuz135/industrial/proxy/BlockRegistry.java
@@ -66,6 +66,7 @@ public class BlockRegistry {
     public static ProteinReactorBlock proteinReactorBlock = new ProteinReactorBlock();
     public static ProteinGeneratorBlock proteinGeneratorBlock = new ProteinGeneratorBlock();
     public static HydratorBlock hydratorBlock = new HydratorBlock();
+    public static WitherBuilderBlock witherBuilderBlock = new WitherBuilderBlock();
 
     public static IFCustomFluidBlock BLOCK_ESSENCE = new IFCustomFluidBlock(FluidsRegistry.ESSENCE, Material.WATER);
     public static IFCustomFluidBlock BLOCK_MILK = new IFCustomFluidBlock(FluidsRegistry.MILK, Material.WATER);