+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 + Fixed [#83](https://github.com/Buuz135/Industrial-Foregoing/issues/83)
 + Added Leaf Shearing Addon for the Plant Gatherer
 + Fixed [#81](https://github.com/Buuz135/Industrial-Foregoing/issues/81)
++ Fixed [#85](https://github.com/Buuz135/Industrial-Foregoing/issues/85)
 
 # 1.5.2
 + Fixed dirt to diamonds