+++ CHANGELOG.md
@@ -1,9 +1,10 @@
-# 1.6.0
+# 1.5.3
 + Added some changes to be able to pass information from the Gatherer to the Recollectable
 + Fixed [#83](https://github.com/Buuz135/Industrial-Foregoing/issues/83)
 + Added Leaf Shearing Addon for the Plant Gatherer
 + Fixed [#81](https://github.com/Buuz135/Industrial-Foregoing/issues/81)
 + Fixed [#85](https://github.com/Buuz135/Industrial-Foregoing/issues/85)
++ Added Wither Builder
 
 # 1.5.2
 + Fixed dirt to diamonds